# native/application/api/openapi/swagger.yaml
basePath: /
definitions:
...
  application_api_controller.SendEmailCaptchaRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  application_api_controller.SendSMSCaptchaRequest:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
...
  github_com_gcc798_lightning_application_api_domain_request.BatchDeleteDictRequest:
    properties:
      ids:
        description: 字典ID列表
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - ids
    type: object
...
  github_com_gcc798_lightning_application_api_domain_request.CreateDictRequest:
    properties:
      createBy:
        description: 创建者
        type: integer
      dictLabel:
        description: 字典标签
        type: string
      dictType:
        description: 字典类型
        type: string
      dictValue:
        description: 字典键值
        type: string
      isDefault:
        description: 是否默认
        type: boolean
      parentId:
        description: 父字典ID（0表示根节点）
        type: integer
      remark:
        description: 备注
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态：0正常 1停用
        type: integer
      updateBy:
        description: 更新者
        type: integer
    required:
    - dictLabel
    - dictType
    - dictValue
    type: object
...
  github_com_gcc798_lightning_application_api_domain_request.UpdateDictRequest:
    properties:
      dictLabel:
        description: 字典标签
        type: string
      dictType:
        description: 字典类型
        type: string
      dictValue:
        description: 字典键值
        type: string
      id:
        description: 字典ID（由路径参数注入）
        type: integer
      isDefault:
        description: 是否默认
        type: boolean
      parentId:
        description: 父字典ID
        type: integer
      remark:
        description: 备注
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态
        type: integer
      updateBy:
        description: 更新者
        type: integer
    required:
    - dictLabel
    - dictType
    - dictValue
    type: object
...
  github_com_gcc798_lightning_application_api_domain_response.DictDataResponse:
    properties:
      createBy:
        description: 创建者
        type: integer
      createdTime:
        description: 创建时间
        type: string
      dictLabel:
        description: 字典标签
        type: string
      dictType:
        description: 字典类型
        type: string
      dictValue:
        description: 字典键值
        type: string
      id:
        description: 字典ID
        type: integer
      isDefault:
        description: 是否默认
        type: boolean
      parentId:
        description: 父字典ID
        type: integer
      remark:
        description: 备注
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态
        type: integer
      updateBy:
        description: 更新者
        type: integer
      updatedTime:
        description: 更新时间
        type: string
    type: object
...
  github_com_gcc798_lightning_application_api_domain_response.Response:
    description: API 统一响应格式
    properties:
      code:
        description: 业务状态码
        example: 200
        type: integer
      data:
        description: 响应数据
        type: object
      msg:
        description: 响应消息
        example: 操作成功
        type: string
    type: object
...
  github_com_gcc798_lightning_internal_platform_captcha.CaptchaData:
    properties:
      data:
        description: 验证码数据
      expireAt:
        description: 过期时间
        type: string
      id:
        description: 验证码ID
        type: string
      type:
        allOf:
        - $ref: '#/definitions/github_com_gcc798_lightning_internal_platform_captcha.CaptchaType'
        description: 验证码类型
    type: object
  github_com_gcc798_lightning_internal_platform_captcha.CaptchaType:
    enum:
    - image
    - sms
    - email
    type: string
    x-enum-comments:
      CaptchaTypeEmail: 邮箱验证码
      CaptchaTypeImage: 图形验证码
      CaptchaTypeSMS: 短信验证码
    x-enum-descriptions:
    - 图形验证码
    - 短信验证码
    - 邮箱验证码
    x-enum-varnames:
    - CaptchaTypeImage
    - CaptchaTypeSMS
    - CaptchaTypeEmail
host: localhost:9009
info:
  contact:
    email: support@example.com
    name: 技术支持
    url: https://example.com/support
  description: Quick Admin RESTful API 接口文档，支持双 Token 认证机制
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms/
  title: NTZ API 文档
  version: "1.0"
paths:
...
  /api/v1/dict:
    post:
      consumes:
      - application/json
      description: 创建新的字典数据，支持树形结构（通过parentId）
      parameters:
      - description: 创建字典请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_request.CreateDictRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 创建字典
      tags:
      - 字典管理
    put:
      consumes:
      - application/json
      description: 更新字典数据
      parameters:
      - description: 更新字典请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_request.UpdateDictRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 更新字典
      tags:
      - 字典管理
  /api/v1/dict/{id}:
    delete:
      consumes:
      - application/json
      description: 删除单个字典数据（如果有子字典则无法删除）
      parameters:
      - description: 字典ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 删除字典
      tags:
      - 字典管理
    get:
      consumes:
      - application/json
      description: 根据字典ID查询字典详情
      parameters:
      - description: 字典ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.DictDataResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 根据ID查询字典
      tags:
      - 字典管理
  /api/v1/dict/batch:
    delete:
      consumes:
      - application/json
      description: 批量删除字典数据（如果有子字典则无法删除）
      parameters:
      - description: 批量删除请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_request.BatchDeleteDictRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 批量删除字典
      tags:
      - 字典管理
  /api/v1/dict/label:
    get:
      consumes:
      - application/json
      description: 根据字典类型和键值获取对应的标签（用于数据展示）
      parameters:
      - description: 字典类型
        example: '"sys_user_sex"'
        in: query
        name: dictType
        required: true
        type: string
      - description: 字典键值
        example: '"0"'
        in: query
        name: dictValue
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      summary: 根据类型和键值获取标签
      tags:
      - 字典管理
  /api/v1/dict/type:
    get:
      consumes:
      - application/json
      description: 根据字典类型获取字典列表，用于前端下拉框等场景。支持获取子字典列表（通过parentId参数）
      parameters:
      - description: 字典类型
        example: '"sys_user_sex"'
        in: query
        name: dictType
        required: true
        type: string
      - description: 父字典ID（可选，用于获取子字典）
        example: 0
        in: query
        name: parentId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.DictDataResponse'
                  type: array
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      summary: 根据类型获取字典列表
      tags:
      - 字典管理
...
  /api/v1/system/dict/page:
    get:
      consumes:
      - application/json
      description: 分页查询字典列表，支持按类型、标签、状态筛选
      parameters:
      - description: 页码
        in: query
        minimum: 1
        name: pageNum
        required: true
        type: integer
      - description: 每页数量
        in: query
        minimum: 1
        name: pageSize
        required: true
        type: integer
      - description: 排序列
        example: '"sort"'
        in: query
        name: orderByColumn
        type: string
      - description: 排序方向
        enum:
        - asc
        - desc
        example: '"asc"'
        in: query
        name: isAsc
        type: string
      - description: 字典类型
        in: query
        name: dictType
        type: string
      - description: 字典标签（模糊查询）
        in: query
        name: dictLabel
        type: string
      - description: 状态：0正常 1停用 -1全部
        enum:
        - 0
        - 1
        - -1
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
      security:
      - Bearer: []
      summary: 分页查询字典列表
      tags:
      - 字典管理
...
  /captcha/email:
    post:
      consumes:
      - application/json
      parameters:
      - description: 邮箱
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/application_api_controller.SendEmailCaptchaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_gcc798_lightning_internal_platform_captcha.CaptchaData'
              type: object
      summary: 发送邮箱验证码
      tags:
      - 验证码
...
  /captcha/sms:
    post:
      consumes:
      - application/json
      parameters:
      - description: 手机号
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/application_api_controller.SendSMSCaptchaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_gcc798_lightning_application_api_domain_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_gcc798_lightning_internal_platform_captcha.CaptchaData'
              type: object
      summary: 发送短信验证码
      tags:
      - 验证码
...
securityDefinitions:
  Bearer:
    description: '格式: "Bearer {access_token}"，AccessToken 从登录接口获取'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...